Micro-Cap-Experiment-Updated/requirements.txt
pandas>=2.2.2
numpy>=2.3.2
yfinance>=0.2.65
matplotlib>=3.8.4
pandas-datareader>=0.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0
fpdf2>=2.8.0
